feat(webpack): move web-specific options into withWeb plugin
