+++ packages/webpack/src/plugins/stats-json-plugin.ts
@@ -2,7 +2,7 @@ import { Compiler, sources } from 'webpack';
 
 export class StatsJsonPlugin {
   apply(compiler: Compiler) {
-    compiler.hooks.emit.tap('angular-cli-stats', (compilation) => {
+    compiler.hooks.emit.tap('StatsJsonPlugin', (compilation) => {
       const data = JSON.stringify(compilation.getStats().toJson('verbose'));
       compilation.assets[`stats.json`] = new sources.RawSource(data);
     });
